requirements.txt
selene >= 2.0.0b10
Appium-Python-Client == 2.6.1
pytest == 7.1.3
pytest-xdist == 2.5.0
allure-pytest == 2.10.0
requests == 2.28.1
pydantic == 1.10.2
python-dotenv == 0.21.0
selenium ~=4.2.0
